copy,refactor: split copyOneImage and copyMultipleImages
As discussed https://github.com/containers/image/pull/1875 it is
suggested to keep the two implementation in different files since there
independent complexity is increasing.
